% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BICcovar.R
\name{BICcovar}
\alias{BICcovar}
\title{BIC with covariates}
\usage{
BICcovar(nw, dataset)
}
\arguments{
\item{nw}{\code{mhnet} network}

\item{dataset}{data frame with attributes given as 'covs' attribute}
}
\value{
real number
}
\description{
Calculates BIC for the given network and data frame.
}
\examples{
data(Gm17403)
data <- Gm17403[,5:7]
covs <- list(Sex = cbind(as.numeric(Gm17403$sex=="M")),
             Diet = cbind(as.numeric(Gm17403$diet=="HF")),
             Batch = cbind(as.numeric(Gm17403$batch==1)))
attr(data, "covs") <- covs

nw.start <- empty.network(c(names(attr(data, "covs")), names(data)))
BICcovar(nw.start, data)
}
\keyword{manip}

