% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BICcovar2}
\alias{BICcovar2}
\title{Yet another BIC with covariates}
\usage{
BICcovar2(nw, dataset)
}
\arguments{
\item{nw}{\code{mhnet} network}

\item{dataset}{A list with covariates (matrix) as an attribute}
}
\value{
real number
}
\description{
Dataset is a list, covartiates are att
}
\keyword{manip}

