% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mhnet-package.R
\docType{data}
\name{Ecsit_network}
\alias{Ecsit_network}
\title{Ecsit Protein Network}
\source{
Diversity Outbred liver protein expression.
}
\usage{
data(Ecsit_network)
}
\description{
Haplotype probabilities, mRNA and protein liver expression of Ecsit, Ndufaf1, Tmem126b (covariates provided as attributes).
}
\keyword{datasets}

