% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mhnet-package.R
\docType{data}
\name{Gm17403}
\alias{Gm17403}
\title{Gm17403 Expression Network}
\source{
Diversity Outbred liver RNA expression.
}
\usage{
data(Gm17403)
}
\description{
Long non-coding RNA Gm17403 and two correlated genes Pan3 and Ctcfl.
}
\keyword{datasets}

