\name{mhnet-package}
\alias{mhnet-package}
\alias{mhnet}
\docType{package}
\title{
Metropolis-Hastings type simulation and inference for Bayesian Networks
}
\description{
Sampling from distribution factorizing according to a directed
    acyclic graph.
}
\details{
\tabular{ll}{
Package: \tab mhnet\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-06\cr
License: \tab GPL-3\cr
}
## An overview of how to use the package, including the most important functions ~~
}
\author{
Petr Simecek

Maintainer: Petr Simecek <simecek@gmail.com>
}
\references{
## Literature or other references for background information ~~
}
\keyword{ package }
\examples{
## simple examples of the most important functions ~~
}
